move protostar fixture to separate file
