+++ tests/integration/_conftest/__init__.py
@@ -1,0 +1,6 @@
+from .protostar_fixture import (
+    ProtostarFixture,
+    GatewayClientTxInterceptor,
+    TransactionRegistry,
+    build_protostar_fixture,
+)
