+++ tests/integration/commands/calculate_contract_address_command_test.py
@@ -1,7 +1,7 @@
 import pytest
 
 from tests.integration.conftest import CreateProtostarProjectFixture
-from tests.integration.protostar_fixture import ProtostarFixture
+from tests.integration._conftest import ProtostarFixture
 
 
 @pytest.fixture(name="protostar")
