+++ tests/integration/commands/invoke_command_test.py
@@ -5,8 +5,7 @@ from tests._conftest.devnet import DevnetFixture
 from tests.conftest import DevnetAccount, SetPrivateKeyEnvVarFixture
 from tests.data.contracts import RUNTIME_ERROR_CONTRACT
 from tests.integration.conftest import CreateProtostarProjectFixture
-from tests.integration.protostar_fixture import ProtostarFixture
-
+from tests.integration._conftest import ProtostarFixture
 from protostar.starknet_gateway.gateway_facade import InputValidationException
