+++ tests/integration/testing_cache/test_cache.py
@@ -2,7 +2,7 @@ import pytest
 
 from protostar.self.cache_io import CacheIO
 from tests.integration.conftest import CreateProtostarProjectFixture
-from tests.integration.protostar_fixture import ProtostarFixture
+from tests.integration._conftest import ProtostarFixture
 from tests.data.tests import *
